pyproject.toml
[project]
name = "measuring-hate-speech-llms"
version = "0.1.0"
description = "Evaluation pipeline for estimating hate-speech severity from LLM responses on the Measuring Hate Speech corpus."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.88.0",
    "datasets>=3.6.0",
    "google-genai>=1.70.0",
    "loguru>=0.7.3",
    "openai>=2.30.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/mhs_llms"]
